refactor: refactors the project code
