--- Lambda/deletar-item/pom.xml
@@ -1,12 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.example</groupId>
-        <artifactId>Lambda</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-    <artifactId>deletar-item</artifactId>
-    <name>Archetype - deletar-item</name>
-    <url>http://maven.apache.org</url>
-</project>
