--- Lambda/hello-world/src/main/java/com/example/HelloWorldHandler.java
@@ -1,17 +1,0 @@
-package com.example;
-
-import com.amazonaws.services.lambda.runtime.Context;
-import com.amazonaws.services.lambda.runtime.RequestHandler;
-import java.util.HashMap;
-import java.util.Map;
-
-public class HelloWorldHandler implements RequestHandler<Map<String, Object>, Map<String, Object>> {
-
-    @Override
-    public Map<String, Object> handleRequest(Map<String, Object> input, Context context) {
-        Map<String, Object> response = new HashMap<>();
-        response.put("statusCode", 200);
-        response.put("body", "Hellow Terraform");
-        return response;
-    }
-}
