+++ lambda/hello-world/src/main/java/com/example/HelloWorldHandler.java
@@ -1,0 +1,14 @@
+package com.example;
+
+import com.amazonaws.services.lambda.runtime.Context;
+import com.amazonaws.services.lambda.runtime.RequestHandler;
+
+import java.util.Map;
+
+public class HelloWorldHandler implements RequestHandler<Map<String, Object>, Map<String, Object>> {
+
+    @Override
+    public Map<String, Object> handleRequest(Map<String, Object> input, Context context) {
+        return ResponseUtils.buildResponse(200, "Hellow Terraform");
+    }
+}
