+++ terraform/modules/environments/dev/.terraform/providers/registry.terraform.io/hashicorp/aws/5.97.0/windows_386/LICENSE.txt
