+++ terraform/modules/environments/dev/dynamo.tf
