+++ terraform/modules/environments/dev/providers.tf
