+++ terraform/modules/environments/prod/providers.tf
