+++ terraform/modules/lambda/outputs.tf
